Project Objectives
Objective of this project is to make a sales pattern analysis and customer analysis which would be leading to more accurate forecasting and quotas. The sales pattern analysis focuses on finding trends within sales data, which can help you better understand your product demand. The goal of customer analytics is to create a single, accurate view of an organization's customer base, which can inform decisions about how to best acquire and retain future customers. It can also identify high-value customers and suggest proactive ways to interact with them and this project, companies identify their most and least performant products, monitor buyer and consumer behavior, and determine which issues impact the business. And also used by e-commerce professionals to improve both sales and the customer experience, but can also be used by inventory managers for fulfillment and by accountants to track costs and profitability.